requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11

# Database
sqlalchemy==2.0.25

# Scheduling
apscheduler==3.10.4

# Configuration
pyyaml==6.0.1

# Notifications
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
responses==0.25.0

# Utilities
python-dateutil==2.8.2
